TEES: seperate running reader, extracting relevant reader output, and processing
